+++ config/routes.rb
@@ -5,6 +5,7 @@ Rails.application.routes.draw do
 
       mount_devise_token_auth_for "User", at: "auth"
 
+      resources :categories, only: %i[index create update destroy]
       resources :transactions, only: %i[index show create update destroy]
       # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html