+++ spec/factories/transactions.rb
@@ -3,10 +3,12 @@
 FactoryBot.define do
   factory :transaction do
     association :user
+    category { association(:category, user: user) }
     amount { 100.0 }
+    currency { "BRL" }
     kind { :income }
     description { "Salary" }
-    category { "Job" }
     occurred_at { Time.zone.now }
+    paid { true }
   end
 end